+++ src/haloray-core/gui/models/simulationStateModel.h
@@ -33,6 +33,7 @@ public:
         AtmosphereEnabled,
         Turbidity,
         GroundAlbedo,
+        GuidesEnabled,
         NUM_COLUMNS
     };