+++ src/haloray-core/gui/viewSettingsWidget.cpp
@@ -24,6 +24,7 @@ ViewSettingsWidget::ViewSettingsWidget(SimulationStateModel *viewModel, QWidget 
     m_mapper->addMapping(m_pitchSlider, SimulationStateModel::CameraPitch);
     m_mapper->addMapping(m_yawSlider, SimulationStateModel::CameraYaw);
     m_mapper->addMapping(m_hideSubHorizonCheckBox, SimulationStateModel::HideSubHorizon);
+    m_mapper->addMapping(m_showGuidesCheckBox, SimulationStateModel::GuidesEnabled);
     m_mapper->toFirst();
 
     connect(m_cameraProjectionComboBox, QOverload<int>::of(&QComboBox::currentIndexChanged), m_mapper, &QDataWidgetMapper::submit, Qt::QueuedConnection);
@@ -31,6 +32,7 @@ ViewSettingsWidget::ViewSettingsWidget(SimulationStateModel *viewModel, QWidget 
     connect(m_pitchSlider, &SliderSpinBox::valueChanged, m_mapper, &QDataWidgetMapper::submit, Qt::QueuedConnection);
     connect(m_yawSlider, &SliderSpinBox::valueChanged, m_mapper, &QDataWidgetMapper::submit, Qt::QueuedConnection);
     connect(m_hideSubHorizonCheckBox, &QCheckBox::stateChanged, m_mapper, &QDataWidgetMapper::submit, Qt::QueuedConnection);
+    connect(m_showGuidesCheckBox, &QCheckBox::stateChanged, m_mapper, &QDataWidgetMapper::submit, Qt::QueuedConnection);
 
     /*
      * It is not possible to map multiple model columns to different properties
@@ -43,7 +45,7 @@ ViewSettingsWidget::ViewSettingsWidget(SimulationStateModel *viewModel, QWidget 
     m_maximumFovMapper->toFirst();
 
     connect(m_brightnessSlider, &SliderSpinBox::valueChanged, this, &ViewSettingsWidget::brightnessChanged);
-    connect(m_lockToLightSource, &QCheckBox::stateChanged, this, &ViewSettingsWidget::lockToLightSource);
+    connect(m_lockToLightSourceCheckBox, &QCheckBox::stateChanged, this, &ViewSettingsWidget::lockToLightSource);
 }
 
 void ViewSettingsWidget::setupUi()
@@ -70,7 +72,9 @@ void ViewSettingsWidget::setupUi()
 
     m_hideSubHorizonCheckBox = new QCheckBox();
 
-    m_lockToLightSource = new QCheckBox();
+    m_lockToLightSourceCheckBox = new QCheckBox();
+
+    m_showGuidesCheckBox = new QCheckBox();
 
     auto layout = new QFormLayout(this->contentWidget());
     layout->addRow(tr("Camera projection"), m_cameraProjectionComboBox);
@@ -79,7 +83,8 @@ void ViewSettingsWidget::setupUi()
     layout->addRow(tr("Yaw"), m_yawSlider);
     layout->addRow(tr("Brightness"), m_brightnessSlider);
     layout->addRow(tr("Hide sub-horizon"), m_hideSubHorizonCheckBox);
-    layout->addRow(tr("Lock to light source"), m_lockToLightSource);
+    layout->addRow(tr("Lock to light source"), m_lockToLightSourceCheckBox);
+    layout->addRow(tr("Show guides"), m_showGuidesCheckBox);
 }
 
 void ViewSettingsWidget::setBrightness(double brightness)