+++ src/haloray-core/simulation/simulationEngine.cpp
@@ -33,7 +33,8 @@ SimulationEngine::SimulationEngine(
       m_cameraLockedToLightSource(false),
       m_multipleScatteringProbability(0.0),
       m_crystalRepository(crystalRepository),
-      m_atmosphere(Atmosphere::createDefaultAtmosphere())
+      m_atmosphere(Atmosphere::createDefaultAtmosphere()),
+      m_guidesEnabled(false)
 {
     initialize();
 }
@@ -94,6 +95,20 @@ void SimulationEngine::setAtmosphere(Atmosphere atmosphere)
     emit atmosphereChanged(m_atmosphere);
 }
 
+bool SimulationEngine::getGuidesEnabled() const
+{
+    return m_guidesEnabled;
+}
+
+void SimulationEngine::setGuidesEnabled(bool newState)
+{
+    if (m_guidesEnabled == newState) return;
+
+    clear();
+    m_guidesEnabled = newState;
+    emit guidesToggled(m_guidesEnabled);
+}
+
 unsigned int SimulationEngine::getOutputTextureHandle() const
 {
     return m_simulationTexture->getHandle();
@@ -102,6 +117,11 @@ unsigned int SimulationEngine::getOutputTextureHandle() const
 unsigned int SimulationEngine::getBackgroundTextureHandle() const
 {
     return m_backgroundTexture->getHandle();
+}
+
+unsigned int SimulationEngine::getGuideTextureHandle() const
+{
+    return m_guideTexture->getHandle();
 }
 
 unsigned int SimulationEngine::getIteration() const
@@ -126,6 +146,24 @@ void SimulationEngine::stop()
 void SimulationEngine::step()
 {
     ++m_iteration;
+
+    if (m_guidesEnabled && m_iteration == 1)
+    {
+        glMemoryBarrier(GL_SHADER_IMAGE_ACCESS_BARRIER_BIT);
+        glBindImageTexture(m_guideTexture->getTextureUnit(), m_guideTexture->getHandle(), 0, GL_FALSE, 0, GL_READ_WRITE, GL_R32F);
+        m_guideShader->bind();
+        glMemoryBarrier(GL_SHADER_IMAGE_ACCESS_BARRIER_BIT);
+
+        m_guideShader->setUniformValue("sun.altitude", degToRad(m_light.altitude));
+        m_guideShader->setUniformValue("sun.diameter", degToRad(m_light.diameter));
+        m_guideShader->setUniformValue("camera.pitch", degToRad(m_camera.pitch));
+        m_guideShader->setUniformValue("camera.yaw", degToRad(m_camera.yaw));
+        m_guideShader->setUniformValue("camera.focalLength", m_camera.getFocalLength());
+        m_guideShader->setUniformValue("camera.projection", m_camera.projection);
+        m_guideShader->setUniformValue("camera.hideSubHorizon", m_camera.hideSubHorizon ? 1 : 0);
+
+        glDispatchCompute(m_outputWidth, m_outputHeight, 1);
+    }
 
     if (m_atmosphere.enabled && m_iteration == 1)
     {
@@ -231,6 +269,10 @@ void SimulationEngine::clear()
 
     glClearTexImage(m_backgroundTexture->getHandle(), 0, GL_RGBA, GL_FLOAT, NULL);
     glBindImageTexture(m_backgroundTexture->getTextureUnit(), m_backgroundTexture->getHandle(), 0, GL_FALSE, 0, GL_READ_WRITE, GL_RGBA32F);
+
+    glClearTexImage(m_guideTexture->getHandle(), 0, GL_RGBA, GL_FLOAT, NULL);
+    glBindImageTexture(m_guideTexture->getTextureUnit(), m_guideTexture->getHandle(), 0, GL_FALSE, 0, GL_READ_WRITE, GL_R32F);
+
     m_iteration = 0;
 }
 
@@ -262,7 +304,7 @@ void SimulationEngine::initialize()
 void SimulationEngine::initializeShaders()
 {
     qInfo("Initializing raytracing shader");
-    m_simulationShader = std::make_unique<QOpenGLShaderProgram>();
+    m_simulationShader = new QOpenGLShaderProgram(this);
     bool raytraceShaderReadSucceeded = m_simulationShader->addCacheableShaderFromSourceFile(QOpenGLShader::ShaderTypeBit::Compute, ":/shaders/raytrace.glsl");
     if (raytraceShaderReadSucceeded == false)
     {
@@ -294,12 +336,30 @@ void SimulationEngine::initializeShaders()
         throw std::runtime_error(m_skyShader->log().toUtf8());
     }
     qInfo("Sky shader program compilation and linking successful");
+
+    qInfo("Initializing guide marking shader");
+    m_guideShader = new QOpenGLShaderProgram(this);
+    bool guideShaderReadSucceeded = m_guideShader->addCacheableShaderFromSourceFile(QOpenGLShader::ShaderTypeBit::Compute, ":/shaders/guide.glsl");
+    if (guideShaderReadSucceeded == false)
+    {
+        qWarning("Reading guide marking shader failed");
+        throw std::runtime_error(m_guideShader->log().toUtf8());
+    }
+    qInfo("Guide marking shader successfully initialized");
+
+    if (m_guideShader->link() == false)
+    {
+        qWarning("Compiling and linking guide marking shader failed");
+        throw std::runtime_error(m_guideShader->log().toUtf8());
+    }
+    qInfo("Guide marking shader program compilation and linking successful");
 }
 
 void SimulationEngine::initializeTextures()
 {
     m_simulationTexture = std::make_unique<OpenGL::Texture>(m_outputWidth, m_outputHeight, 0, OpenGL::TextureType::Color);
-    m_backgroundTexture = std::make_unique<OpenGL::Texture>(m_outputWidth, m_outputHeight, 2, OpenGL::TextureType::Color);
+    m_backgroundTexture = std::make_unique<OpenGL::Texture>(m_outputWidth, m_outputHeight, 1, OpenGL::TextureType::Color);
+    m_guideTexture = std::make_unique<OpenGL::Texture>(m_outputWidth, m_outputHeight, 2, OpenGL::TextureType::Monochrome);
 }
 
 void SimulationEngine::resizeOutputTextureCallback(const unsigned int width, const unsigned int height)
@@ -309,6 +369,7 @@ void SimulationEngine::resizeOutputTextureCallback(const unsigned int width, con
 
     m_simulationTexture.reset();
     m_backgroundTexture.reset();
+    m_guideTexture.reset();
 
     initializeTextures();
     clear();